pyproject.toml
[project]
name = "digital-soul"
version = "0.1.0"
description = "A Multi-Modal AI system to predict personality from voice."
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "librosa>=0.10.0",
    "moviepy>=1.0.3",
    "numpy<2.0.0", 
    "openai-whisper==20231117",
    "numba==0.58.1",
    "llvmlite==0.41.1",
    "pandas>=2.0.0",
    "soundfile>=0.12.1",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "torchvision>=0.17.0",
    "tqdm>=4.65.0",
    "transformers>=4.48.0",
    "gradio>=4.0.0",
    "matplotlib>=3.7.0"
]

[tool.uv]
package = false
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "pytorch_cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch_cu118", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" }
]
torchvision = [
    { index = "pytorch_cu118", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" }
]
torchaudio = [
    { index = "pytorch_cu118", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" }
]
